// ui/lib/api/schemas/page.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 */
import type { PageId } from './pageId'
import type { PageNodes } from './pageNodes'

export interface Page {
  /** @minimum 0 */
  height: number
  id: PageId
  name: string
  /**
   * Stacking = insertion order. Bottom-first: `source` is typically first,
   * `rendered` typically last.
   */
  nodes: PageNodes
  /** @minimum 0 */
  width: number
}